Made some attempts at iteration 1
I tried a few different versions of the obs and act spaces. I think maybe I should just let the models train for longer at a time. All of the info needed for plotting like rewards and episode lengths etc. are stored in pkl files so rerunning isn't necessary if you want to make changes to the plotting